First Dockerfile version for frontend (#114)
* fix: errors in build fixed

* fix: dropdown button icon removed

* chore: custom palette variables included

* feat: first version of Dockerfile

* fix: theme reestructurated to reduce build problems

* fix: chip colors overriding to fix all build problems

* chore: backend api data commented for now in nginx configuration

* feat: Dynamic env variable injection

* fix: fixes after PR sonar review

* fix: useless env variable removed

* fix: nginx configuration comments removed

* fix: usage of npm as package manager

* fix  --ignore-scripts added to avoid sonar security hotspot

* fix: omit dev flag removed

* chore: version included to nginx image

* fix: environment variable injection modified
