+++ ichub-frontend/.dockerignore
@@ -1,0 +1,6 @@
+.chart
+.git
+.github
+.gitignore
+.husky
+Dockerfile
