+++ ichub-frontend/index.html
@@ -34,6 +34,13 @@
   </head>
   <body>
     <div id="root"></div>
+    <script>
+      // Do NOT change ENV attributes without changing them in scripts/inject-dynamic-env.sh as well
+      const ENV = {
+        REQUIRE_HTTPS_URL_PATTERN: "true",
+        ICHUB_BACKEND_URL: "https://portal-backend.example.org"
+      }
+    </script>
     <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
