+++ ichub-frontend/package.json
@@ -7,7 +7,9 @@
     "dev": "vite",
     "build": "tsc -b && vite build",
     "lint": "eslint .",
-    "preview": "vite preview"
+    "preview": "vite preview",    
+    "build:docker": "IMAGE=ichub-frontend && docker build -t $IMAGE -f Dockerfile .",
+    "start:docker": "IMAGE=ichub-frontend && docker run --rm -d -p 8080:80 --name ichub-frontend $IMAGE"
   },
   "dependencies": {
     "@catena-x/portal-shared-components": "^3.7.6",
