+++ ichub-frontend/src/components/general/CardChip.tsx
@@ -82,7 +82,7 @@ export const CardChip = ({ status, statusText }: CardChipProps) => {
         border: theme.palette.chip[border],
         height: '28px',
       }}
-      icon={statusKey==StatusVariants.shared?<PersonIcon sx={{color: '#000000', fontSize: '18px'}}/>:null}
+      icon={statusKey==StatusVariants.shared?<PersonIcon sx={{color: '#000000', fontSize: '18px'}}/>:undefined}
     />
   )
 }
