+++ ichub-frontend/src/theme/palette.ts
@@ -1,0 +1,178 @@
+/********************************************************************************
+ * Copyright (c) 2023 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Apache License, Version 2.0 which is available at
+ * https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ********************************************************************************/
+
+export const paletteDefinitions = {
+  common: {
+    white: '#fff',
+    black: '#000',
+  },
+  primary: {
+    main: 'rgb(1,32,96)',
+    dark: 'black',
+    contrastText: '#fff',
+    shadow: 'rgba(15, 113, 203, 0.4)',
+  },
+  secondary: {
+    main: '#eaf1fe',
+    dark: '#d4e3fe',
+    contrastText: '#0f71cb',
+  },
+  brand: {
+    brand01: '#FFA600',
+    brand02: '#B3CB2D',
+  },
+  action: {
+    active: '#939393',
+    disabled: '#ADADAD',
+    disabledBackground: '#EAEAEA',
+  },
+  danger: {
+    danger: '#D91E18',
+    dangerHover: '#E5231D',
+    dangerBadge: '#FB6540',
+  },
+  support: {
+    success: '#00AA55',
+    error: '#D91E18',
+    info: '#676BC6',
+    warning: '#FFA600',
+  },
+  success: {
+    main: '#00aa55',
+  },
+  icon: {
+    icon01: '#939393',
+    icon02: '#B6B6B6',
+    icon03: '#333333',
+  },
+  border: {
+    border01: '#DCDCDC',
+    border02: '#B6B6B6',
+    border03: '#989898',
+    border04: '#0F71CB',
+  },
+  chip: {
+    release: '#0D61AE',
+    active: '#88982D',
+    inactive: '#D91E18',
+    created: '#8f8013',
+    inReview: '#8f8013',
+    enabled: '#007877',
+    default: '#888888',
+    bgRelease: '#EAF1FE',
+    bgActive: '#F0F5D5',
+    bgInactive: '#FFF7FF',
+    bgCreated: '#f5efd5',
+    bgInReview: '#f5efd5',
+    bgEnabled: '#DEEEEF',
+    bgDefault: '#c7c5c5',
+    warning: '#FFA600',
+    registered: 'rgb(255, 255, 255)',
+    bgRegistered: '2px solid rgba(40, 104, 255, 1)',
+    borderDraft: '1px solid rgba(246, 246, 246, 0.66)',
+    black: "#000000",
+    none: 'none'
+  },
+  stepper: {
+    stepDone: '#B3CB2D',
+    stepCurrent: '#FFA600',
+    stepUpcoming: '#EAEAEA',
+  },
+  background: {
+    background01: '#F9F9F9',
+    background02: '#F3F3F3',
+    background03: '#E9E9E9',
+    background04: '#F4FBFD',
+    background05: '#F5F9EE',
+    background06: '#FFF7EC',
+    background07: '#F5F5F5',
+    background08: '#FFF6FF',
+    background09: '#EDF0F4',
+    background10: '#303030F2',
+    background11: '#EDEFF2',
+    background12: '#8E8E8E1A',
+    background13: '#DFE4EA',
+    background14: '#303030',
+  },
+  textField: {
+    placeholderText: '#8D8D8D',
+    helperText: '#717171',
+    background: '#F7F7F7',
+    backgroundHover: '#ECECEC',
+  },
+  text: {
+    primary: '#111111',
+    secondary: '#252525',
+    tertiary: '#888888',
+    quaternary: '#A2A2A2',
+  },
+  accent: {
+    accent01: '#4D73D5',
+    accent02: '#F2F3FB',
+    accent03: '#676BC6',
+    accent04: '#E1F1FF',
+    accent05: '#FFEBCC',
+    accent06: '#5E3416',
+    accent07: '#88982D',
+    accent08: '#F0F5D5',
+    accent09: '#FDB943',
+    accent10: '#428C5B',
+    accent11: '#337B89',
+    accent12: '#2B4078',
+  },
+  selected: {
+    hover: 'rgba(15, 113, 203, 0.05)',
+    focus: 'rgba(15, 113, 203, 0.15)',
+    active: 'rgba(15, 113, 203, 0.2)',
+  },
+  pending: {
+    main: '#FFECBD',
+    contrastText: '#975B27',
+  },
+  confirmed: {
+    main: '#ffffff',
+    contrastText: '#ffffff',
+  },
+  declined: {
+    main: '#FEE7E2',
+    contrastText: '#FF532F',
+  },
+  label: {
+    main: '#F2F3FB',
+    contrastText: '#676BC6',
+  },
+  deleted: {
+    main: '#eaeaea',
+    contrastText: '#adadad',
+  },
+  info: {
+    main: '#F2F3FB',
+    contrastText: '#676BC6',
+  },
+  warning: {
+    main: '#ffa602',
+    contrastText: '#ffff',
+  },
+  buttons: {
+    darkGrey: '#e1e1e1',
+    lightGrey: '#f3f3f3',
+    white: '#f9f9f9',
+    yellow: '#f5f9ee',
+  },
+}
